% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ArrangeData}
\alias{ArrangeData}
\title{combine different data sets togather}
\usage{
ArrangeData(type)
}
\arguments{
\item{type}{a string which can be "seedling", "tissue" or "leaf"}
}
\value{
a list
\item{data} a n-by-p data frame of counts, with row name being the gene IDs
\item{trt} a p-vector indicating the treatment factor
\item{group} a p-vector of experimental group IDs
}
\description{
Combine the different GSE data sets to make Set 1 , Set 2 and Set 3.
}

