% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catchToList}
\alias{catchToList}
\title{Poisson regression with random effect}
\usage{
catchToList(expr)
}
\arguments{
\item{expr}{a regression expression}
}
\value{
a list
\item{value} a list returned by the regression model
\item{warnings} warning message. If none, "NO" is returned.
\item{error} error message. If none, "NO" is returned.
}
\description{
run poisson regression with random effect , and keep track of warnings
  or errors of regression for a single gene
}

