% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.geneColumn}
\alias{create.geneColumn}
\title{Create gene IDs as a column}
\usage{
create.geneColumn(data)
}
\arguments{
\item{data}{input data}
}
\value{
return a new data set
}
\description{
Converting a data.frame whose row.names are gene into  a new data
containing the row.names as a new column, facilitating to merge by gene
}

