% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate.var.component}
\alias{estimate.var.component}
\title{estimate variance component for all genes piecewisely.}
\usage{
estimate.var.component(data, group, trt, filter.factor = 3)
}
\arguments{
\item{data}{an n-by-p data frame that combines experimental data to be analyzed, and row names are gene IDs}

\item{group}{a p-vector of ids,  indicating to which experiment the column should belong}

\item{trt}{a p-vector, treatment ids nested in group ids.}
}
\value{
a list of two data frames. One is the read count matrix, and the other
is a data frame described below.
an n-by-7 data frame containing the columns
\item{gene} gene names
\item{individual} variance component for biological sample
\item{trt} variance component for treatment
\item{lab} variance component for experiment
\item{warnS} a 1-0 indicator, 1 if there is a warning and 0 if not.
\item{sum}  the sum of individual, trt and lab
\item{Rank} the rank of sum (column 6) in ascending order
}
\description{
regression for all genes within a dataset.
}

