% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match.gene}
\alias{match.gene}
\title{Match the top ranked genes}
\usage{
match.gene(genelist, var.comp, top)
}
\arguments{
\item{genelist}{a data frame containing the top gene IDs}

\item{var.comp}{ranked genes by sum of variance components.}

\item{top}{how many genes are to be matched}
}
\value{
a list
\item{s} the number of genes matched
\item{Gene} the gene list successfully matched
\item{Rank} the corresponding ranks
}
\description{
match the top ranked genes
}

