% Generated by roxygen2 (4.0.2): do not edit by hand
\name{norm.factor}
\alias{norm.factor}
\title{Estimate norm. factor}
\usage{
norm.factor(var.comp, data, n)
}
\arguments{
\item{var.comp}{ranked genes by sum of variance components.}

\item{data}{data frame of read count}

\item{n}{number of top stably expressed genes}
}
\value{
a list
\item{s} the number of genes matched
\item{Gene} the gene list successfully matched
\item{Rank} the corresponding ranks
}
\description{
estimate norm factors by most stably expressed genes
}

