% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.gene}
\alias{plot.gene}
\title{Plot CPM of a gene list}
\usage{
\method{plot}{gene}(genelist, count, lower = 1, upper = 10000, figure.num)
}
\arguments{
\item{genelist}{a list of gene IDs}

\item{lower}{lower limit for the y-axis}

\item{upper}{upper limit for the y-axis}

\item{figure.num}{A string that is used to name the plot}

\item{data}{read count data set to be plotted}
}
\value{
a figure
}
\description{
Plot the Counts Per Million (CPM) of a given gene list from RNA-Seq data
}

