% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.rank}
\alias{plot.rank}
\title{Plot the rank of genes}
\usage{
\method{plot}{rank}(genelist, var.comp, col = "blue", figure.num)
}
\arguments{
\item{genelist}{a data frame containing the gene IDs}

\item{var.comp}{ranked genes by sum of variance components.}

\item{col}{the color of the dot}

\item{figure.num}{A string that is used to name the plot}
}
\value{
a figure
}
\description{
Plot the rank of a given gene list
}

