% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SRAtoFASTQ}
\alias{SRAtoFASTQ}
\title{downloading SRA file and unzip it to FASTQ}
\usage{
SRAtoFASTQ(file, SAR.path)
}
\arguments{
\item{file}{A RunInfo file (.csv format) generated from NCBI, which should be in current working directory.}

\item{SRA.path}{where the SRA Toolkit "bin" folder locates}

\item{the}{path where FASTQ file should be stored}
}
\value{
the returned objects will be used in \code{FASTQtoCount} function
}
\description{
Download SRA file from NCBI and convert it into FASTQ
}
\details{
SRA files can be downloaded automatically from NCBI, and then are converted into "FASTQ" file
using SRA Toolkit (version 2.3.5-2)
}
\note{
need to install SRA toolkit first. In the "RunInfo" file, choose \code{LibraryStrategy=="RNA-Seq"},
\code{LibrarySelection=="cDNA"}, and \code{RunInfo$LibraryLayout=="SINGLE"}.
}

